Little bit refactoring to take CheckedFunctional out.
